test(server): Test invalid RunWf request with Basic Workflow
